music-api/requirements.txt
urllib3<=2.0.7
ytmusicapi
pyyaml